requirements.txt
streamlit==1.34.0
Faker==19.2.0
hugchat==0.4.6
sqlalchemy==2.0.30
psycopg2-binary
